contracts: reindent

+++ hardhat/contracts/dbanta.sol
@@ -51,12 +51,11 @@ library SafeMath {
     }
 }
 
-
 contract Dbanta {
     using SafeMath for uint256;
     address public owner; //Owner is also a maintainer
     bool public stopped = false;
-    uint public bantCount = 0;
+    uint256 public bantCount = 0;
 
     struct User {
         uint256 id;
@@ -78,7 +77,7 @@ contract Dbanta {
         uint256 timestamp;
         uint256 likeCount;
         uint256 reportCount;
-        uint tipVote;
+        uint256 tipVote;
         cdStatus status; // Bant Active-Deleted-Banned
     }
 
@@ -89,16 +88,24 @@ contract Dbanta {
         string content;
         uint256 likeCount;
         uint256 timestamp;
-        cdStatus status; 
+        cdStatus status;
     }
 
     uint256 public totalBants = 0;
     uint256 public totalComments = 0;
     uint256 public totalUsers = 0;
 
-    enum accountStatus {Active, inActive, Deactivated}
+    enum accountStatus {
+        Active,
+        inActive,
+        Deactivated
+    }
 
-    enum cdStatus {Active, inActive, Deleted } 
+    enum cdStatus {
+        Active,
+        inActive,
+        Deleted
+    }
 
     //Comment-Bant status
     // enum BantStatus{NP,Active, Banned, Deleted}
@@ -170,10 +177,20 @@ contract Dbanta {
 
     event logRegisterUser(address user, uint256 id);
     event logUserBanned(address user, uint256 id);
-    event logBantCreated(address payable author, uint256 userid, uint256 bantid, string hashtag, uint tipVote);
+    event logBantCreated(
+        address payable author,
+        uint256 userid,
+        uint256 bantid,
+        string hashtag,
+        uint256 tipVote
+    );
     event logBantDeleted(uint256 id, string hashtag);
-    event bantVoted(address payable author, uint256 userid, uint256 bantid, uint tipVote);
-
+    event bantVoted(
+        address payable author,
+        uint256 userid,
+        uint256 bantid,
+        uint256 tipVote
+    );
 
     constructor() {
         owner = msg.sender;
@@ -184,7 +201,7 @@ contract Dbanta {
         revert();
     }
 
-//User Details
+    //User Details
     function usernameAvailable(string memory _username)
         public
         view
@@ -212,7 +229,7 @@ contract Dbanta {
         checkUserNotExists(msg.sender)
         usernameTaken(_username)
     {
-        usernames[_username] = true; 
+        usernames[_username] = true;
         totalUsers = totalUsers.add(1);
         uint256 id = totalUsers;
         users[msg.sender] = User(
@@ -310,7 +327,6 @@ contract Dbanta {
         );
     }
 
- 
     function createBant(
         string memory _hashtag,
         string memory _content,
@@ -319,7 +335,7 @@ contract Dbanta {
         totalBants = totalBants.add(1);
         uint256 id = totalBants;
         require(bytes(_content).length > 0);
-        bantCount ++;
+        bantCount++;
         Bants[id] = Bant(
             id,
             payable(msg.sender),
@@ -331,7 +347,6 @@ contract Dbanta {
             0,
             0,
             cdStatus.Active
-            
         );
         userBants[msg.sender].push(totalBants);
         emit logBantCreated(
@@ -343,12 +358,12 @@ contract Dbanta {
         );
     }
 
-    function tipPost(uint _id) public payable {
+    function tipPost(uint256 _id) public payable {
         // Make sure the id is valid
         require(_id > 0 && _id <= bantCount);
         // Fetch the post
         Bant memory _bant = Bants[_id];
-        User memory _user = users [msg.sender];
+        User memory _user = users[msg.sender];
         // Fetch the author
         address payable _author = _bant.author;
         // Pay the author by sending them tokens
@@ -356,10 +371,10 @@ contract Dbanta {
         // Incremet the tip amount
         _bant.tipVote = _bant.tipVote + msg.value;
         // Update the post
-       Bants[_id] = _bant;
+        Bants[_id] = _bant;
         // Trigger an event
         //address payable author, uint256 userid, uint256 bantid, uint tipVote
-        emit bantVoted(_author, _bant.bantId,_user.id, _bant.tipVote);
+        emit bantVoted(_author, _bant.bantId, _user.id, _bant.tipVote);
     }
 
     /// @notice Edit a bant
@@ -468,7 +483,6 @@ contract Dbanta {
     {
         return userBants[_user];
     }
-
 
     /// @notice Create a comment on bant
     /// @param  _bantid Id of bantList
@@ -584,9 +598,6 @@ contract Dbanta {
     {
         return (bantComments[_id]);
     }
-
-   
-    
 
     /*
      ****************************************Owner Admin ******************************************************************************************
